ai-server/vbd/requirements.txt
accelerate==1.7.0
aioboto3==14.3.0
aiobotocore==2.22.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.7
aioitertools==0.12.0
aiosignal==1.3.2
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
astroid==3.3.9
attrs==25.3.0
bcrypt==4.3.0
betterpath==0.2.2
boto3==1.37.3
botocore==1.37.3
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
diffusers==0.32.2
dill==0.4.0
Django==5.1.7
django-cors-headers==4.7.0
djangorestframework==3.15.2
einops
ecdsa==0.19.1
fastapi==0.115.12
fastapi_nextauth_jwt==2.0.1
filelock==3.18.0
frozenlist==1.6.0
fsspec==2025.3.0
graphviz==0.20.3
greenlet==3.1.1
h11==0.16.0
hkdf==0.0.3
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.29.3
idna==3.10
imageio==2.37.0
importlib_metadata==8.6.1
iniconfig==2.1.0
ip-adapter @ git+https://github.com/tencent-ailab/IP-Adapter.git
isort==6.0.1
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.1
jwt==1.3.1
l==0.11.0
lazy_loader==0.4
MarkupSafe==3.0.2
mccabe==0.7.0
mpmath==1.3.0
multidict==6.4.4
networkx==3.5
nibabel==5.3.2
numpy==2.2.4
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
packaging==24.2
passlib==1.7.4
pillow==11.1.0
platformdirs==4.3.7
pluggy==1.5.0
propcache==0.3.1
psutil==7.0.0
py2puml==0.10.0
pyasn1==0.4.8
pycparser==2.22
pydantic==2.11.0
pydantic-settings==2.9.1
pydantic_core==2.33.0
PyJWT==2.10.1
pylint==3.3.6
pytest==8.3.5
pytest-asyncio==0.26.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rsa==4.9.1
s3transfer==0.11.3
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.3
setuptools==78.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
sqlparse==0.5.3
starlette==0.46.1
sympy==1.14.0
threadpoolctl==3.6.0
tifffile==2025.6.1
tomlkit==0.13.2
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
triton==3.3.1
transformers==4.48.2
typing-inspection==0.4.0
typing_extensions==4.13.0
uml-class-diagram-generator==0.1
urllib3==2.3.0
uvicorn==0.34.0
vcversioner==2.16.0.0
wrapt==1.17.2
yarl==1.20.0
zipp==3.21.0
zope.interface==7.2
